requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-cors-headers==2.5.2
django-extensions==2.1.6
django-timezone-field==3.0
djangorestframework==3.9.2
dnspython==1.16.0
eventlet==0.24.1
greenlet==0.4.15
idna==2.8
monotonic==1.5
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
sqlparse==0.3.0
urllib3==1.24.1
vine==1.3.0
psycopg2==2.8.2